Rename bottom-bar to bottom-bar-container
[#173471252]
