+++ src/geohazard-components/bottom-bar-container.scss
