+++ src/geohazard-components/bottom-bar-container.tsx
@@ -2,7 +2,7 @@ import React, { useEffect, useState } from "react";
 import CCLogo from "./assets/cc-logo.svg";
 import CCLogoSmall from "./assets/cc-logo-small.svg";
 import screenfull from "screenfull";
-import css from "./bottom-bar.scss";
+import css from "./bottom-bar-container.scss";
 
 const toggleFullscreen = () => {
   if (!screenfull || !screenfull.isEnabled) {
